Levels 7-8 validation: multi-file projects, regression + determinism gates (50/50)
Scenarios 46-48 exercise aura.toml projects (multi-file root, path
dependency lib, larger app); 49 is a regression corpus trapping every
bug found by the ladder; 50 is the determinism gate hashing a fixed
workload in both engines.

Found by 49: integer literals were stored as u64 — values past
u64::MAX silently clamped to u64::MAX. Widened Literal::Int and
Const::Int to u128 (parser → MIR → codegen iconcat → interp), and
parse_int_lit now emits E0004 on u128 overflow instead of a silent
clamp. Permanent regressions in aura-parser and aura-interp tests.

+++ crates/aura-ast/src/lib.rs
@@ -136,7 +136,7 @@ impl Ast {
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Literal {
-    Int(u64),
+    Int(u128),
     Float(f64),
     Bool(bool),
     /// `()`
